backend/requirements.txt
Django>=6,<6.1
django-ninja>=1,<2
django-allauth[headless]>=65,<66
django-cors-headers>=4,<5
inngest>=0.5,<0.6
requests>=2.32,<3
psycopg[binary]>=3.2,<4
agno>=1.8,<2
openai>=1.99,<2
